requirements.txt
certifi==2019.3.9
chardet==3.0.4
configparser==3.7.4
future==0.17.1
idna==2.8
mysql-connector-python==8.0.15
oauthlib==3.0.1
pika==1.0.0
protobuf==3.7.0
python-twitter==3.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
